+++ src/components/pages/deprecatedFrom31/AnimalAgreement/AnimalYes.js
