+++ src/components/pages/deprecatedFrom31/ClientRelease.js
@@ -5,7 +5,7 @@ Requires Signatures from Client and Staff for clause in Client Release of inform
 import React from 'react';
 
 //Previous/Next buttons
-import IntakeButton from '../../IntakeButtons';
+import IntakeButton from '../IntakePacketContent/IntakeButtons';
 
 //Ant Design imports (https://ant.design/components/overview/)
 import { Card, Progress } from 'antd';
