+++ src/components/pages/deprecatedFrom31/CoreValues.js
