+++ src/components/pages/deprecatedFrom31/ThirdPartyConsent/ThirdPartySigs.js
@@ -5,7 +5,7 @@ Requires Signatures from Client and Staff for clause in Third Party Consent form
 import React from 'react';
 
 //Previous/Next buttons
-import IntakeButton from '../../IntakeButtons';
+import IntakeButton from '../../IntakePacketContent/IntakeButtons';
 
 //Ant Design imports (https://ant.design/components/overview/)
 import { Form, Input, Card, Progress, DatePicker, Row, Col } from 'antd';
